fetch. 프로필 수정 관련 코드 완성

+++ src/main/java/com/zero/triptalk/like/entity/PlannerLike.java
@@ -23,7 +23,7 @@ public class PlannerLike {
     @JoinColumn(name = "planner_id")
     private Planner planner;
 
-    private Double likeCount;
+    private Long likeCount;
 
     private LocalDateTime likeDt;
 }
